align governance section surface
