+++ docs/authorization-inventory.md
@@ -119,10 +119,6 @@ The `Source` column reflects the constant referenced in the attribute as it appe
 | `OnboardingReviewController.Clear` | Action | `ConsentCoordinator, Board, Admin` | `PolicyNames.ConsentCoordinatorBoardOrAdmin` |
 | `OnboardingReviewController.Flag` | Action | `ConsentCoordinator, Board, Admin` | `PolicyNames.ConsentCoordinatorBoardOrAdmin` |
 | `OnboardingReviewController.Reject` | Action | `ConsentCoordinator, Board, Admin` | `PolicyNames.ConsentCoordinatorBoardOrAdmin` |
-| `OnboardingReviewController.BoardVoting` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
-| `OnboardingReviewController.BoardVotingDetail` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
-| `OnboardingReviewController.Vote` | Action | `Board` | `PolicyNames.BoardOnly` |
-| `OnboardingReviewController.Finalize` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
 
 ### Governance / Application Section
 
@@ -130,9 +126,12 @@ The `Source` column reflects the constant referenced in the attribute as it appe
 |---|---|---|---|
 | `GovernanceController` | Class | `[Authorize]` (authenticated) | — |
 | `GovernanceController.Roles` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
-| `ApplicationController` | Class | `[Authorize]` (authenticated) | — |
-| `ApplicationController.Applications` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
-| `ApplicationController.ApplicationDetail` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
+| `GovernanceApplicationsController` | Class | `[Authorize]` (authenticated) | — |
+| `GovernanceApplicationsController.Admin` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
+| `GovernanceApplicationsController.AdminDetail` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
+| `GovernanceBoardVotingController` | Class | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
+| `GovernanceBoardVotingController.Vote` | Action | `Board` | `PolicyNames.BoardOnly` |
+| `GovernanceBoardVotingController.Finalize` | Action | `Board, Admin` | `PolicyNames.BoardOrAdmin` |
 
 ### Profile / Contacts Section
 
@@ -336,8 +335,8 @@ Views express authorization four ways today:
 
 | View | Line | Check | Controls |
 |---|---|---|---|
-| `OnboardingReview/BoardVotingDetail.cshtml` | 115 | `authorize-policy="BoardOnly"` | Vote casting card |
-| `OnboardingReview/BoardVotingDetail.cshtml` | 153 | `Model.CanFinalize` | Finalize decision card |
+| `Governance/BoardVoting/Detail.cshtml` | 115 | `authorize-policy="BoardOnly"` | Vote casting card |
+| `Governance/BoardVoting/Detail.cshtml` | 153 | `Model.CanFinalize` | Finalize decision card |
 | `OnboardingReview/Detail.cshtml` | 30 | `Model.HasPendingApplication` | Application tier badge |
 
 ### Team Views
