+++ docs/guide/Onboarding.md
@@ -85,7 +85,7 @@ If you hold the **Volunteer Coordinator** role, you have read-only access to the
 Board members and Admins can do everything a Consent Coordinator can, plus:
 
 - **Resolve flagged profiles.** Only Board or Admin can manually approve a flagged human (the Approve button on `/Profile/{id}/Admin`).
-- **Vote on tier applications.** Board votes on Colaborador / Asociado applications at `/OnboardingReview/BoardVoting`; Admin can finalize a vote, set the meeting date, and override.
+- **Vote on tier applications.** Board votes on Colaborador / Asociado applications at `/Governance/BoardVoting`; Admin can finalize a vote, set the meeting date, and override.
 - **Review the full onboarding pipeline.** See [humans](Glossary.md#human) at every stage, including those stuck waiting on documents or a coordinator.
 
 Admins and all coordinator roles bypass the membership gate entirely (`MembershipRequiredFilter` consults `RoleChecks.BypassesMembershipRequirement`), so you reach the full app regardless of your own onboarding status. Suspended Admins/Board keep their role claims so they can manage their own unsuspension, but lose the `ActiveMember` claim.
