+++ src/Humans.Application/DTOs/Governance/BoardVotingDetailData.cs
@@ -4,9 +4,9 @@ using NodaTime;
 namespace Humans.Application.DTOs.Governance;
 
 /// <summary>
-/// Detail projection for the Board Voting detail view
-/// (<c>BoardVotingDetail.cshtml</c>). Replaces the old <c>Application</c>
-/// return from <c>OnboardingService.GetBoardVotingDetailAsync</c> whose
+/// Detail projection for the Governance Board Voting detail view
+/// (<c>Views/Governance/BoardVoting/Detail.cshtml</c>). Replaces the old <c>Application</c>
+/// return from the pre-migration board-voting flow whose
 /// <c>.User.Profile</c> and <c>.BoardVotes[].BoardMemberUser</c> chains have
 /// been stitched via <c>IUserService</c> + <c>IProfileService</c>.
 /// </summary>
