+++ src/Humans.Web/Controllers/OnboardingReviewController.cs
@@ -2,9 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
-using NodaTime;
 using Humans.Application.DTOs;
-using Humans.Application.Interfaces.Governance;
 using Humans.Domain.Constants;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
@@ -24,7 +22,6 @@ namespace Humans.Web.Controllers;
 public class OnboardingReviewController : HumansControllerBase
 {
     private readonly IOnboardingService _onboardingService;
-    private readonly IApplicationDecisionService _applicationDecisionService;
     private readonly IUserService _userService;
     private readonly ILogger<OnboardingReviewController> _logger;
     private readonly IStringLocalizer<SharedResource> _localizer;
@@ -32,14 +29,12 @@ public class OnboardingReviewController : HumansControllerBase
     public OnboardingReviewController(
         UserManager<User> userManager,
         IOnboardingService onboardingService,
-        IApplicationDecisionService applicationDecisionService,
         IUserService userService,
         ILogger<OnboardingReviewController> logger,
         IStringLocalizer<SharedResource> localizer)
         : base(userManager)
     {
         _onboardingService = onboardingService;
-        _applicationDecisionService = applicationDecisionService;
         _userService = userService;
         _logger = logger;
         _localizer = localizer;
@@ -236,210 +231,6 @@ public class OnboardingReviewController : HumansControllerBase
             SetError(_localizer["Common_Error"].Value);
         }
         return RedirectToAction(nameof(Index));
-    }
-
-    [HttpGet("BoardVoting")]
-    [Authorize(Policy = PolicyNames.BoardOrAdmin)]
-    public async Task<IActionResult> BoardVoting(CancellationToken ct)
-    {
-        var (applications, boardMembers) = await _applicationDecisionService.GetBoardVotingDashboardAsync(ct);
-
-        var viewModel = new BoardVotingDashboardViewModel
-        {
-            BoardMembers = boardMembers
-                .Select(m => new BoardVoteMemberViewModel
-                {
-                    UserId = m.UserId,
-                    DisplayName = m.DisplayName
-                })
-                .ToList(),
-            Applications = applications.Select(a =>
-            {
-                var appVm = new BoardVotingApplicationViewModel
-                {
-                    ApplicationId = a.ApplicationId,
-                    UserId = a.UserId,
-                    DisplayName = a.UserDisplayName,
-                    ProfilePictureUrl = a.UserProfilePictureUrl,
-                    MembershipTier = a.MembershipTier,
-                    ApplicationMotivation = a.ApplicationMotivation,
-                    SubmittedAt = a.SubmittedAt.ToDateTimeUtc(),
-                    Status = a.Status
-                };
-                foreach (var vote in a.Votes)
-                {
-                    appVm.VotesByBoardMember[vote.BoardMemberUserId] = new BoardVoteCellViewModel
-                    {
-                        Vote = vote.Vote,
-                        Note = vote.Note
-                    };
-                }
-                return appVm;
-            }).ToList(),
-        };
-
-        return View(viewModel);
-    }
-
-    [HttpGet("BoardVoting/{applicationId:guid}")]
-    [Authorize(Policy = PolicyNames.BoardOrAdmin)]
-    public async Task<IActionResult> BoardVotingDetail(Guid applicationId, CancellationToken ct)
-    {
-        var application = await _applicationDecisionService.GetBoardVotingDetailAsync(applicationId, ct);
-        if (application is null)
-            return NotFound();
-
-        var currentUser = await GetCurrentUserAsync();
-        if (currentUser is null)
-            return NotFound();
-
-        var currentVote = application.Votes.FirstOrDefault(v => v.BoardMemberUserId == currentUser.Id);
-        var isAdmin = RoleChecks.IsAdmin(User);
-
-        var viewModel = new BoardVotingDetailViewModel
-        {
-            ApplicationId = application.ApplicationId,
-            UserId = application.UserId,
-            DisplayName = application.DisplayName,
-            ProfilePictureUrl = application.ProfilePictureUrl,
-            Email = application.Email,
-            FirstName = application.FirstName,
-            LastName = application.LastName,
-            City = application.City,
-            CountryCode = application.CountryCode,
-            MembershipTier = application.MembershipTier,
-            Status = application.Status,
-            ApplicationMotivation = application.Motivation,
-            AdditionalInfo = application.AdditionalInfo,
-            SignificantContribution = application.SignificantContribution,
-            RoleUnderstanding = application.RoleUnderstanding,
-            SubmittedAt = application.SubmittedAt.ToDateTimeUtc(),
-            Votes = application.Votes
-                .Select(v => new BoardVoteDetailItemViewModel
-                {
-                    BoardMemberUserId = v.BoardMemberUserId,
-                    DisplayName = v.BoardMemberDisplayName ?? string.Empty,
-                    Vote = v.Vote,
-                    Note = v.Note,
-                    VotedAt = v.VotedAt.ToDateTimeUtc()
-                })
-                .OrderBy(v => v.DisplayName, StringComparer.OrdinalIgnoreCase)
-                .ToList(),
-            CurrentUserVote = currentVote?.Vote,
-            CurrentUserNote = currentVote?.Note,
-            CanFinalize = isAdmin
-        };
-
-        return View(viewModel);
-    }
-
-    [HttpPost("BoardVoting/Vote")]
-    [ValidateAntiForgeryToken]
-    [Authorize(Policy = PolicyNames.BoardOnly)]
-    public async Task<IActionResult> Vote(Guid applicationId, VoteChoice vote, string? note)
-    {
-        var currentUser = await GetCurrentUserAsync();
-        if (currentUser is null)
-            return NotFound();
-
-        try
-        {
-            var result = await _applicationDecisionService.CastBoardVoteAsync(
-                applicationId, currentUser.Id, vote, note);
-
-            if (!result.Success)
-            {
-                SetError(result.ErrorKey switch
-                {
-                    "NotFound" => _localizer["BoardVoting_ApplicationNotFound"].Value,
-                    "NotSubmitted" => _localizer["BoardVoting_ApplicationNotVotable"].Value,
-                    _ => _localizer["BoardVoting_ApplicationNotVotable"].Value
-                });
-                return RedirectToAction(nameof(BoardVoting));
-            }
-
-            SetSuccess(_localizer["BoardVoting_VoteSaved"].Value);
-            return RedirectToAction(nameof(BoardVotingDetail), new { applicationId });
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to cast board vote for application {ApplicationId}", applicationId);
-            SetError(_localizer["BoardVoting_ApplicationNotVotable"].Value);
-            return RedirectToAction(nameof(BoardVoting));
-        }
-    }
-
-    [HttpPost("BoardVoting/Finalize")]
-    [ValidateAntiForgeryToken]
-    [Authorize(Policy = PolicyNames.BoardOrAdmin)]
-    public async Task<IActionResult> Finalize(BoardVotingFinalizeModel model)
-    {
-        var currentUser = await GetCurrentUserAsync();
-        if (currentUser is null)
-            return NotFound();
-
-        // Require a valid meeting date (validated in controller as it's form input)
-        LocalDate? meetingDate = null;
-        if (!string.IsNullOrWhiteSpace(model.BoardMeetingDate))
-        {
-            var pattern = NodaTime.Text.LocalDatePattern.Iso;
-            var parseResult = pattern.Parse(model.BoardMeetingDate);
-            if (parseResult.Success)
-                meetingDate = parseResult.Value;
-        }
-
-        if (meetingDate is null)
-        {
-            SetError(_localizer["BoardVoting_MeetingDateRequired"].Value);
-            return RedirectToAction(nameof(BoardVotingDetail), new { applicationId = model.ApplicationId });
-        }
-
-        // Require at least one board vote before finalization
-        var hasVotes = await _applicationDecisionService.HasBoardVotesAsync(model.ApplicationId);
-        if (!hasVotes)
-        {
-            SetError(_localizer["BoardVoting_NoVotes"].Value);
-            return RedirectToAction(nameof(BoardVotingDetail), new { applicationId = model.ApplicationId });
-        }
-
-        try
-        {
-            ApplicationDecisionResult result;
-            if (model.Approved)
-            {
-                result = await _applicationDecisionService.ApproveAsync(
-                    model.ApplicationId, currentUser.Id,
-                    model.DecisionNote, meetingDate);
-            }
-            else
-            {
-                result = await _applicationDecisionService.RejectAsync(
-                    model.ApplicationId, currentUser.Id,
-                    model.DecisionNote ?? string.Empty, meetingDate);
-            }
-
-            if (!result.Success)
-            {
-                _logger.LogWarning("Finalize failed for application {ApplicationId}: {ErrorKey}",
-                    model.ApplicationId, result.ErrorKey);
-                SetError(result.ErrorKey switch
-                {
-                    "NotFound" => _localizer["BoardVoting_ApplicationNotFound"].Value,
-                    "NotSubmitted" => _localizer["BoardVoting_ApplicationNotVotable"].Value,
-                    "ConcurrencyConflict" => _localizer["BoardVoting_ConcurrencyConflict"].Value,
-                    _ => _localizer["BoardVoting_ApplicationNotVotable"].Value
-                });
-                return RedirectToAction(nameof(BoardVoting));
-            }
-
-            SetSuccess(_localizer["BoardVoting_Finalized"].Value);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to finalize application {ApplicationId}", model.ApplicationId);
-            SetError(_localizer["BoardVoting_ApplicationNotVotable"].Value);
-        }
-        return RedirectToAction(nameof(BoardVoting));
     }
 
     private static OnboardingReviewItemViewModel MapToItem(
