+++ src/Humans.Web/Extensions/Sections/UsersSectionExtensions.cs
@@ -1,14 +1,11 @@
 using Humans.Application.Interfaces.Dashboard;
 using Humans.Application.Interfaces.Gdpr;
-using Humans.Application.Interfaces.Governance;
 using Humans.Application.Interfaces.Repositories;
 using Humans.Application.Interfaces.Users;
 using Humans.Application.Services.Users.AccountLifecycle;
 using Humans.Infrastructure.Repositories.Users;
 using DashboardAdminDashboardService = Humans.Application.Services.Dashboard.AdminDashboardService;
 using DashboardDashboardService = Humans.Application.Services.Dashboard.DashboardService;
-using GovernanceMembershipCalculator = Humans.Application.Services.Governance.MembershipCalculator;
-using GovernanceMembershipQuery = Humans.Application.Services.Governance.MembershipQuery;
 using UsersUserService = Humans.Application.Services.Users.UserService;
 
 namespace Humans.Web.Extensions.Sections;
@@ -36,12 +33,6 @@ internal static class UsersSectionExtensions
         // no outbound edges to higher-level sections.
         services.AddScoped<IAccountDeletionService, AccountDeletionService>();
 
-        // Query adapter breaks the circular DI graph between IMembershipCalculator
-        // and ITeamService / IRoleAssignmentService (both of which inject
-        // ISystemTeamSync, whose implementation injects IMembershipCalculator back).
-        // Only MembershipCalculator depends on the query adapter.
-        services.AddScoped<IMembershipQuery, GovernanceMembershipQuery>();
-        services.AddScoped<IMembershipCalculator, GovernanceMembershipCalculator>();
         services.AddScoped<IDashboardService, DashboardDashboardService>();
         // Admin dashboard aggregator — owns no tables; aggregates user
         // partition, application stats, and language distribution from
