+++ src/Humans.Web/Resources/SharedResource.es.resx
@@ -1089,7 +1089,7 @@
     &lt;li&gt;&lt;strong&gt;Solicitante:&lt;/strong&gt; {0}&lt;/li&gt;
     &lt;li&gt;&lt;strong&gt;ID de solicitud:&lt;/strong&gt; {1}&lt;/li&gt;
 &lt;/ul&gt;
-&lt;p&gt;&lt;a href="{2}/Board/Applications"&gt;Revisar solicitud&lt;/a&gt;&lt;/p&gt;</value><comment>{0} = applicant name, {1} = application ID, {2} = base URL</comment></data>
+&lt;p&gt;&lt;a href="{2}/Governance/Applications/Admin/{1}"&gt;Revisar solicitud&lt;/a&gt;&lt;/p&gt;</value><comment>{0} = applicant name, {1} = application ID, {2} = base URL</comment></data>
 
   <data name="Email_ApplicationApproved_Body" xml:space="preserve"><value>&lt;h2&gt;&#xA1;Tu solicitud ha sido aprobada!&lt;/h2&gt;
 &lt;p&gt;Hola {0},&lt;/p&gt;
@@ -1195,7 +1195,7 @@
 &lt;p&gt;Hola {0},&lt;/p&gt;
 &lt;p&gt;Tu mandato de &lt;strong&gt;{1}&lt;/strong&gt; expira el &lt;strong&gt;{2}&lt;/strong&gt;.&lt;/p&gt;
 &lt;p&gt;Si deseas continuar como {1}, env&#xED;a una solicitud de renovaci&#xF3;n antes de que finalice tu mandato.&lt;/p&gt;
-&lt;p&gt;&lt;a href="{3}/Application/Create"&gt;Enviar solicitud de renovaci&#xF3;n&lt;/a&gt;&lt;/p&gt;
+&lt;p&gt;&lt;a href="{3}/Governance/Applications/Create"&gt;Enviar solicitud de renovaci&#xF3;n&lt;/a&gt;&lt;/p&gt;
 &lt;p&gt;Si no renuevas, seguir&#xE1;s siendo un human activo pero ya no tendr&#xE1;s acceso de nivel {1}.&lt;/p&gt;
 &lt;p&gt;&#xA1;Gracias por tu participaci&#xF3;n!&lt;br/&gt;El equipo de Humans&lt;/p&gt;</value><comment>{0} = user name, {1} = tier name, {2} = expiry date, {3} = base URL</comment></data>
