+++ src/Humans.Web/Views/Governance/Applications/Create.cshtml
@@ -5,7 +5,7 @@
 
 <nav aria-label="breadcrumb">
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a asp-action="Index">@Localizer["Governance_Title"]</a></li>
+        <li class="breadcrumb-item"><a asp-controller="Governance" asp-action="Index">@Localizer["Governance_Title"]</a></li>
         <li class="breadcrumb-item active" aria-current="page">@Localizer["GovernanceCreate_NewApplication"]</li>
     </ol>
 </nav>
