+++ src/Humans.Web/Views/Governance/BoardVoting/Index.cshtml
@@ -5,8 +5,8 @@
 
 <div class="d-flex justify-content-between align-items-center mb-4">
     <h1>@Localizer["BoardVoting_Title"]</h1>
-    <a asp-action="Index" class="btn btn-outline-secondary">
-        <i class="fa-solid fa-arrow-left me-1"></i> @Localizer["OnboardingReview_Title"]
+    <a asp-controller="Governance" asp-action="Index" class="btn btn-outline-secondary">
+        <i class="fa-solid fa-arrow-left me-1"></i> @Localizer["Governance_Title"]
     </a>
 </div>
