+++ src/Humans.Web/Views/Shared/_ApplicationsListContent.cshtml
@@ -15,27 +15,27 @@
 <div class="card mb-4">
     <div class="card-body d-flex flex-wrap gap-3 align-items-center">
         <div class="btn-group" role="group">
-            <a asp-action="Applications" asp-route-tier="@Model.TierFilter" class="btn @(string.IsNullOrEmpty(Model.StatusFilter) ? "btn-primary" : "btn-outline-primary")">
+            <a asp-action="Admin" asp-route-tier="@Model.TierFilter" class="btn @(string.IsNullOrEmpty(Model.StatusFilter) ? "btn-primary" : "btn-outline-primary")">
                 @Localizer["AdminApp_Pending"]
             </a>
-            <a asp-action="Applications" asp-route-status="Approved" asp-route-tier="@Model.TierFilter" class="btn @(Model.StatusFilter == "Approved" ? "btn-primary" : "btn-outline-primary")">
+            <a asp-action="Admin" asp-route-status="Approved" asp-route-tier="@Model.TierFilter" class="btn @(Model.StatusFilter == "Approved" ? "btn-primary" : "btn-outline-primary")">
                 @Localizer["AdminHuman_Approved"]
             </a>
-            <a asp-action="Applications" asp-route-status="Rejected" asp-route-tier="@Model.TierFilter" class="btn @(Model.StatusFilter == "Rejected" ? "btn-primary" : "btn-outline-primary")">
+            <a asp-action="Admin" asp-route-status="Rejected" asp-route-tier="@Model.TierFilter" class="btn @(Model.StatusFilter == "Rejected" ? "btn-primary" : "btn-outline-primary")">
                 @Localizer["AdminApp_Rejected"]
             </a>
-            <a asp-action="Applications" asp-route-status="Withdrawn" asp-route-tier="@Model.TierFilter" class="btn @(Model.StatusFilter == "Withdrawn" ? "btn-primary" : "btn-outline-primary")">
+            <a asp-action="Admin" asp-route-status="Withdrawn" asp-route-tier="@Model.TierFilter" class="btn @(Model.StatusFilter == "Withdrawn" ? "btn-primary" : "btn-outline-primary")">
                 @Localizer["AdminApp_Withdrawn"]
             </a>
         </div>
         <div class="btn-group" role="group">
-            <a asp-action="Applications" asp-route-status="@Model.StatusFilter" class="btn @(string.IsNullOrEmpty(Model.TierFilter) ? "btn-secondary" : "btn-outline-secondary") btn-sm">
+            <a asp-action="Admin" asp-route-status="@Model.StatusFilter" class="btn @(string.IsNullOrEmpty(Model.TierFilter) ? "btn-secondary" : "btn-outline-secondary") btn-sm">
                 All Tiers
             </a>
-            <a asp-action="Applications" asp-route-status="@Model.StatusFilter" asp-route-tier="Colaborador" class="btn @(Model.TierFilter == "Colaborador" ? "btn-secondary" : "btn-outline-secondary") btn-sm">
+            <a asp-action="Admin" asp-route-status="@Model.StatusFilter" asp-route-tier="Colaborador" class="btn @(Model.TierFilter == "Colaborador" ? "btn-secondary" : "btn-outline-secondary") btn-sm">
                 Colaborador
             </a>
-            <a asp-action="Applications" asp-route-status="@Model.StatusFilter" asp-route-tier="Asociado" class="btn @(Model.TierFilter == "Asociado" ? "btn-secondary" : "btn-outline-secondary") btn-sm">
+            <a asp-action="Admin" asp-route-status="@Model.StatusFilter" asp-route-tier="Asociado" class="btn @(Model.TierFilter == "Asociado" ? "btn-secondary" : "btn-outline-secondary") btn-sm">
                 Asociado
             </a>
         </div>
@@ -78,7 +78,7 @@
                                 <small class="text-muted">@app.MotivationPreview</small>
                             </td>
                             <td class="text-end">
-                                <a asp-action="ApplicationDetail" asp-route-id="@app.Id" class="btn btn-sm btn-outline-primary">
+                                <a asp-action="AdminDetail" asp-route-id="@app.Id" class="btn btn-sm btn-outline-primary">
                                     @Localizer["AdminApp_Review"]
                                 </a>
                             </td>
@@ -97,7 +97,7 @@
             @for (var i = 1; i <= Model.TotalPages && i <= 10; i++)
             {
                 <li class="page-item @(i == Model.PageNumber ? "active" : "")">
-                    <a class="page-link" asp-action="Applications" asp-route-page="@i" asp-route-status="@Model.StatusFilter" asp-route-tier="@Model.TierFilter">@i</a>
+                    <a class="page-link" asp-action="Admin" asp-route-page="@i" asp-route-status="@Model.StatusFilter" asp-route-tier="@Model.TierFilter">@i</a>
                 </li>
             }
         </ul>
